refactor context menu
